Extract common deployment code
