+++ lambdas/notify_old_deploys/deployment_utils.py
@@ -1,0 +1,1 @@
+../common/deployment_utils.py
