+++ lambdas/notify_old_deploys/ecs_utils.py
@@ -1,0 +1,1 @@
+../common/ecs_utils.py
