+++ lambdas/service_deployment_status/service_deployment_status.py
@@ -7,70 +7,15 @@ This lambda tracks deployment status from the ECS Task State change event
 stream and writes deployment "color" to dynamo. With color being blue or green.
 """
 
-import collections
-from datetime import datetime
-import dateutil.parser
 import os
 import pprint
 
 import boto3
 
-from ecs_utils import get_cluster_arns, get_service_arns, describe_service, describe_cluster
+from deployment_utils import get_deployments_from_dynamo, get_deployments_from_ecs
 
 
-Deployment = collections.namedtuple('Deployment', 'deployment_key deployment_status color created_at task_definition')
-DeploymentKey = collections.namedtuple('Deployment', 'id service_arn')
-
-def _create_deployment_tuple_from_ecs(service, deployment):
-    deployment_status = deployment['status']
-    ongoing_deployment = len(service['deployments']) > 1
-
-    color = "blue"
-    if(ongoing_deployment):
-        color = "green" if(deployment_status == "PRIMARY") else "blue"
-
-    return Deployment(
-        DeploymentKey(deployment['id'], service['serviceArn']),
-        deployment_status,
-        color,
-        deployment['createdAt'],
-        deployment['taskDefinition']
-    )
-
-def _get_service_deployments(ecs_client, cluster_arn, service_arn):
-    service = describe_service(ecs_client, cluster_arn, service_arn)
-
-    return [_create_deployment_tuple_from_ecs(service, deployment) for deployment in service['deployments']]
-
-def _create_deployment_tuple_from_item(item):
-    return Deployment(
-        DeploymentKey(item['deployment_id'],item['service_arn']),
-        item['deployment_status'],
-        item['color'],
-        dateutil.parser.parse(item['created_at'],""),
-        item['task_definition']
-    )
-
-
-def get_deployments_from_dynamo(table):
-    response = table.scan()
-
-    return [_create_deployment_tuple_from_item(d) for d in response['Items']]
-
-
-def get_deployments_from_ecs(ecs_client):
-    deployments=[]
-
-    cluster_arns = get_cluster_arns(ecs_client)
-    for cluster_arn in cluster_arns:
-        service_arns = get_service_arns(ecs_client, cluster_arn)
-        for service_arn in service_arns:
-            service_deployments = _get_service_deployments(ecs_client, cluster_arn, service_arn)
-            deployments += service_deployments
-
-    return deployments
-
-def find_in_deployments(deployment_list, key):
+def _find_in_deployments(deployment_list, key):
     return [deployment for deployment in deployment_list if deployment[0] == key][0]
 
 def delete_deployment_in_dynamo(table, deployment):
@@ -123,9 +68,9 @@ def compare_deployments(current_deployments, last_deployments):
 
     unchanged_deployments = set(current_deployments) & set(last_deployments)
 
-    deleted_deployments = [find_in_deployments(last_deployments, deletion_key) for deletion_key in deletions_keys]
-    added_deployments = [find_in_deployments(current_deployments, additions_key) for additions_key in additions_keys]
-    maybe_updated_deployments = [find_in_deployments(current_deployments, remaining_key) for remaining_key in remaining_keys]
+    deleted_deployments = [_find_in_deployments(last_deployments, deletion_key) for deletion_key in deletions_keys]
+    added_deployments = [_find_in_deployments(current_deployments, additions_key) for additions_key in additions_keys]
+    maybe_updated_deployments = [_find_in_deployments(current_deployments, remaining_key) for remaining_key in remaining_keys]
 
     updated_deployments = list(set(maybe_updated_deployments) - unchanged_deployments)
